+++ app/src/main/java/com/lagradost/cloudstream3/ui/player/FullScreenPlayer.kt
@@ -255,7 +255,7 @@ open class FullScreenPlayer : AbstractPlayerFragment() {
         playerBinding?.bottomPlayerBar?.post {
             @OptIn(UnstableApi::class)
             val sView = subView ?: return@post
-            val sStyle = subStyle ?: return@post
+            val sStyle = CustomDecoder.style
             val binding = playerBinding ?: return@post
 
             val move = if (isShowing) minOf(
